Rename module to fix import issues.
