+++ deb/openmediavault/usr/lib/python3/dist-packages/openmediavault/device/plugins/md.py
@@ -21,7 +21,7 @@
 import os
 import re
 
-import openmediavault.collections
+import openmediavault.collectiontools
 import openmediavault.device
 import openmediavault.subprocess
